requirements.txt
cufflinks==0.17.3
keras==2.9.0
matplotlib==3.5.2
mplfinance==0.12.9b1
numpy==1.22.4
pandas==1.4.2
pandas_datareader==0.10.0
plotly==5.8.2
python_dateutil==2.8.2
scikit_learn==1.1.1
seaborn==0.11.2
streamlit==1.10.0
